+++ server/encode.h
@@ -14,15 +14,15 @@
 */
 
 // These guys do ascii encoding / decoding of binary data in src.
-bool asciiEncode(const char* src, const int srclen, char** dst);
-bool asciiDecode(const char* src, int* dstlen, char** dst);
+bool asciiEncode(const char* src, const size_t srclen, char** dst);
+bool asciiDecode(const char* src, size_t* dstlen, char** dst);
 
 // These guys do hex-ascii encoding / decoding of binary data in src.
-bool hexEncode(const char* src, const int srclen, char** dst);
-bool hexDecode(const char* src, int* dstlen, char** dst);
+bool hexEncode(const char* src, const size_t srclen, char** dst);
+bool hexDecode(const char* src, size_t* dstlen, char** dst);
 
 // This detects whether the current message encoding is suitable for direct
 // display on a terminal.
-bool isAscii(const char* data, const int len);
+bool isAscii(const char* data, const size_t len);
 
 #endif
